Remove mixins from Display
